refactor: merge QuorumObserver and ObserverContext classes to the one entity to reduce amount of abstractions
