+++ src/llmq/observer/context.cpp
@@ -6,9 +6,11 @@
 
 #include <llmq/debug.h>
 #include <llmq/dkgsessionmgr.h>
-#include <llmq/observer/quorums.h>
-#include <llmq/quorumsman.h>
+#include <llmq/quorums.h>
+#include <msg_result.h>
 
+#include <chain.h>
+#include <net.h>
 #include <validation.h>
 
 namespace llmq {
@@ -18,18 +20,16 @@ ObserverContext::ObserverContext(CBLSWorker& bls_worker, CConnman& connman, CDet
                                  llmq::CQuorumSnapshotManager& qsnapman, const ChainstateManager& chainman,
                                  const CSporkManager& sporkman, const llmq::QvvecSyncModeMap& sync_map,
                                  const util::DbWrapperParams& db_params, bool quorums_recovery) :
-    m_qman{qman},
+    QuorumRoleBase{connman, dmnman, qman, qsnapman, chainman, mn_sync, sporkman, sync_map, quorums_recovery},
     dkgdbgman{std::make_unique<llmq::CDKGDebugManager>(dmnman, qsnapman, chainman)},
     qdkgsman{std::make_unique<llmq::CDKGSessionManager>(dmnman, qsnapman, chainman, sporkman, db_params,
-                                                        /*quorums_watch=*/true)},
-    qman_handler{std::make_unique<llmq::QuorumObserver>(connman, dmnman, qman, qsnapman, chainman, mn_sync, sporkman,
-                                                        sync_map, quorums_recovery)}
+                                                        /*quorums_watch=*/true)}
 {
     qdkgsman->InitializeHandlers([&](const Consensus::LLMQParams& llmq_params,
                                      [[maybe_unused]] int quorum_idx) -> std::unique_ptr<llmq::CDKGSessionHandler> {
         return std::make_unique<llmq::CDKGSessionHandler>(llmq_params);
     });
-    m_qman.ConnectManagers(qman_handler.get(), qdkgsman.get());
+    m_qman.ConnectManagers(this, qdkgsman.get());
 }
 
 ObserverContext::~ObserverContext()
@@ -37,14 +37,23 @@ ObserverContext::~ObserverContext()
     m_qman.DisconnectManagers();
 }
 
-void ObserverContext::Start(int16_t worker_count)
+MessageProcessingResult ObserverContext::ProcessContribQGETDATA(bool request_limit_exceeded, CDataStream& vStream,
+                                                                const CQuorum& quorum, CQuorumDataRequest& request,
+                                                                gsl::not_null<const CBlockIndex*> block_index)
 {
-    qman_handler->Start(worker_count);
+    // Watch-only nodes cannot provide encrypted contributions
+    if (request.GetDataMask() & CQuorumDataRequest::ENCRYPTED_CONTRIBUTIONS) {
+        request.SetError(CQuorumDataRequest::Errors::ENCRYPTED_CONTRIBUTIONS_MISSING);
+        return request_limit_exceeded ? MisbehavingError{25, "request limit exceeded"} : MessageProcessingResult{};
+    }
+    return {};
 }
 
-void ObserverContext::Stop()
+MessageProcessingResult ObserverContext::ProcessContribQDATA(CNode& pfrom, CDataStream& vStream, CQuorum& quorum,
+                                                             CQuorumDataRequest& request)
 {
-    qman_handler->Stop();
+    // Watch-only nodes ignore encrypted contributions
+    return {};
 }
 
 void ObserverContext::InitializeCurrentBlockTip(const CBlockIndex* tip, bool ibd)
@@ -61,6 +70,6 @@ void ObserverContext::UpdatedBlockTip(const CBlockIndex* pindexNew, const CBlock
         return;
 
     qdkgsman->UpdatedBlockTip(pindexNew, fInitialDownload);
-    qman_handler->UpdatedBlockTip(pindexNew, fInitialDownload);
+    QuorumRoleBase::UpdatedBlockTip(pindexNew, fInitialDownload);
 }
 } // namespace llmq
