+++ src/llmq/observer/context.h
@@ -5,37 +5,39 @@
 #ifndef BITCOIN_LLMQ_OBSERVER_CONTEXT_H
 #define BITCOIN_LLMQ_OBSERVER_CONTEXT_H
 
-#include <llmq/options.h>
+#include <llmq/quorumsman.h>
 
 #include <validationinterface.h>
+
+#include <gsl/pointers.h>
+#include <span.h>
 
 #include <memory>
 
 class CBLSWorker;
 class CBlockIndex;
 class CConnman;
+class CDataStream;
 class CDeterministicMNManager;
-class ChainstateManager;
 class CMasternodeMetaMan;
 class CMasternodeSync;
+class CNode;
 class CSporkManager;
+struct MessageProcessingResult;
 namespace llmq {
 class CDKGDebugManager;
 class CDKGSessionManager;
+class CQuorum;
 class CQuorumBlockProcessor;
-class CQuorumManager;
+class CQuorumDataRequest;
 class CQuorumSnapshotManager;
-class QuorumObserver;
 } // namespace llmq
 namespace util {
 struct DbWrapperParams;
 } // namespace util
 
 namespace llmq {
-struct ObserverContext final : public CValidationInterface {
-private:
-    llmq::CQuorumManager& m_qman;
-
+struct ObserverContext final : public QuorumRoleBase, public QuorumRole, public CValidationInterface {
 public:
     ObserverContext() = delete;
     ObserverContext(const ObserverContext&) = delete;
@@ -47,10 +49,19 @@ public:
                     const util::DbWrapperParams& db_params, bool quorums_recovery);
     ~ObserverContext();
 
-    void Start(int16_t worker_count);
-    void Stop();
     void InitializeCurrentBlockTip(const CBlockIndex* tip, bool ibd);
 
+    // QuorumRole
+    // Watch-only nodes are not masternodes
+    bool IsMasternode() const override { return false; }
+    // We are only initialized if watch-only mode is enabled
+    bool IsWatching() const override { return true; }
+    bool SetQuorumSecretKeyShare(CQuorum& quorum, Span<CBLSSecretKey> skContributions) const override { return false; }
+    [[nodiscard]] MessageProcessingResult ProcessContribQGETDATA(bool request_limit_exceeded, CDataStream& vStream,
+                                                                 const CQuorum& quorum, CQuorumDataRequest& request,
+                                                                 gsl::not_null<const CBlockIndex*> block_index) override;
+    [[nodiscard]] MessageProcessingResult ProcessContribQDATA(CNode& pfrom, CDataStream& vStream, CQuorum& quorum,
+                                                              CQuorumDataRequest& request) override;
 protected:
     // CValidationInterface
     void UpdatedBlockTip(const CBlockIndex* pindexNew, const CBlockIndex* pindexFork, bool fInitialDownload) override;
@@ -58,9 +69,6 @@ protected:
 public:
     const std::unique_ptr<llmq::CDKGDebugManager> dkgdbgman;
     const std::unique_ptr<llmq::CDKGSessionManager> qdkgsman;
-
-private:
-    const std::unique_ptr<llmq::QuorumObserver> qman_handler;
 };
 } // namespace llmq
